backend_app/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
sentence-transformers
numpy
google-generativeai
langchain-text-splitters
pydantic
psycopg2-binary